requirements/development.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
asyncio==3.4.3
attrs==24.1.0
certifi==2024.7.4
colorama==0.4.6
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
icecream==2.1.3
idna==3.7
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
yarl==1.9.4
